vet/requirements.txt
streamlit==1.41.0
pip==23.3.1
google-generativeai==0.3.2
protobuf==4.25.3
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0